refactor(models): standardize model names to singular convention and update relations
- Rename Ideas and Rooms models to singular Idea and Room.
- Update HasMany and BelongsTo relationships in Author, Room, and Idea models.
- Fix foreign key references in create_ideas_table migration to point to Room model.
- Change type column definition in create_authors_table migration to unsignedSmallInteger with status documentation.

+++ database/migrations/2026_08_26_144750_create_authors_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name',100);
             $table->text('avatar')->nullable();
-            $table->tinyInteger('type')->default(0);
+            $table->unsignedSmallInteger('type')->default(0); // 0 = Human, 1 = AI
             $table->timestamps();
         });
     }
